mktplinkfw2: add support to pack bootloader
The 841v13 and 841v14 can not flash any image without bootloader over
the WebUI due a bug in the oem firmware upgrade software.

Allow mktplinkfw2 to integrate bootloaders.
An image with the bootloader looks:

|------------|
|image header|
|------------|
|bootloader  |
|------------|
|optional pad|
|------------|
|image header|
|------------|
|kernel image|
|------------|
|rootfs image|
|------------|

The padding depends on the bootloader size. The 2nd image header must
start on a 64k boundary.

Signed-off-by: Alexander Couzens <lynxis@fe80.eu>