tplink-safeloader: TP-Link RE650 v2 support
Support creating images for TP-Link RE650 v2.

Partition table pulled from OEM firmware on the device:
  partition fs-uboot base 0x00000 size 0x20000
  partition os-image base 0x20000 size 0x330000
  partition file-system base 0x350000 size 0x470000
  partition partition-table base 0x7c0000 size 0x02000
  partition default-mac base 0x7c2000 size 0x00020
  partition pin base 0x7c2100 size 0x00020
  partition product-info base 0x7c3100 size 0x01000
  partition soft-version base 0x7c4200 size 0x01000
  partition support-list base 0x7c5200 size 0x01000
  partition profile base 0x7c6200 size 0x08000
  partition config-info base 0x7ce200 size 0x00400
  partition user-config base 0x7d0000 size 0x10000
  partition default-config base 0x7e0000 size 0x10000
  partition radio base 0x7f0000 size 0x10000

One notable difference from v1 in terms of safeloader, is that it requires
soft_ver to be defined for device to accept the firmware, otherwise it
otherwise it denies the firmware file. Version v1 has also 16MB of Flash
instead of 8MB that v2 has, so the paritions had to be shrinked
accordingly.

In OpenWRT "firmware" partition is merged into one, it's made out of
os-image and file-system partitions.

Signed-off-by: Marcin Gordziejewski <openwrt@flicksfix.com>