tplink-safeloader: don't end-pad sysupgrade file
To ensure the stock rootfs ends up at the correct offset, the preceding
kernel partition is padded with 0xff, corresponding to erased flash.
Since on sysupgrade all the required flash space is anyway rased before
writing the new image, it is not necessary to also pad after the second
and last part of the generated image. This can reduce the generated
image size by several MiB.

Signed-off-by: Sander Vanheule <sander@svanheule.net>