ptgen: allow non-default placement of gpt entry table
Usually, the primary GPT entries places in the sector 2 (next sector after
the primary GPT header). The boot ROM of some SoCs (like Airoha AN7581) on
the other hand reads the bootloader code from only one fixed offset which
overlaps with the normal position of the primary GPT entries.

This patch adds a parameter to ptgen to customise the position of the primary
GPT entries.

Signed-off-by: Mikhail Kshevetskiy <mikhail.kshevetskiy@iopsys.eu>