ptgen: do not create stub partition to fill a gap if gap caused by alignment
ptgen tends to create stub partitions between the end of GPT entry table
and the start of the first disk. Normally fdisk/gdisk aligns partition
to 1MB boundary, so there is 2014 sectors gap before the first partition.
It is absolutely normal.

This patch somehow mimics fdisk/gdisk behavior and does NOT create stub
partition in the gap caused by alignment.

Signed-off-by: Mikhail Kshevetskiy <mikhail.kshevetskiy@iopsys.eu>