dkmgt-fwutil: New tool to manipulate TP-Link Omada firmware files
This adds a new tool to create and manipulate TP-Link Omada firmware
files. This was created mostly by reverse engineering the firmware
and bootloader on an the TP-Link ER8411. It also seems to be
compatible with the firmware for the ER605 and possibly other models
too.

Although the firmware contains an RSA signature, the bootloader doesn't
seem to verify it, meaning that this can be used to create factory
installation images via the bootloader's recovery mode.

Signed-off-by: Naomi Kirby <dev@manawolf.ca>